+++ weaver/wps_restapi/processes/processes.py
@@ -31,7 +31,7 @@ from weaver.processes.execution import submit_job, submit_job_dispatch_wps
 from weaver.processes.utils import deploy_process_from_payload, get_process, update_process_metadata
 from weaver.status import Status
 from weaver.store.base import StoreJobs, StoreProcesses
-from weaver.utils import clean_json_text_body, fully_qualified_name, get_any_id, get_header
+from weaver.utils import clean_json_text_body, fully_qualified_name, get_any_id, get_header, make_link_header
 from weaver.visibility import Visibility
 from weaver.wps_restapi import swagger_definitions as sd
 from weaver.wps_restapi.processes.utils import get_process_list_links, get_processes_filtered_by_valid_schemas
@@ -156,6 +156,10 @@ def get_processes(request):
         body["description"] = sd.OkGetProcessesListResponse.description
         LOGGER.debug("Process listing generated, validating schema...")
         body = sd.MultiProcessesListing().deserialize(body)
+        request.response.headers.extend([
+            ("Link", make_link_header(link))
+            for link in body["links"]
+        ])
         return Box(body)
 
     except ServiceException as exc:
@@ -248,13 +252,13 @@ def patch_local_process(request):
 @sd.process_service.get(
     tags=[sd.TAG_PROCESSES, sd.TAG_DESCRIBEPROCESS],
     schema=sd.ProcessEndpoint(),
-    accept=ContentType.TEXT_XML,
+    accept=ContentType.ANY_XML,
     response_schemas=sd.get_process_responses,
 )
 @sd.process_service.get(
     tags=[sd.TAG_PROCESSES, sd.TAG_DESCRIBEPROCESS],
     schema=sd.ProcessEndpoint(),
-    accept=ContentType.APP_XML,
+    accept=ContentType.APP_YAML,
     response_schemas=sd.get_process_responses,
 )
 @sd.process_service.get(
@@ -276,23 +280,48 @@ def get_local_process(request):
         schema = request.params.get("schema")
         ctype = guess_target_format(request)
         ctype_json = add_content_type_charset(ContentType.APP_JSON, "UTF-8")
+        ctype_yaml = add_content_type_charset(ContentType.APP_YAML, "UTF-8")
         ctype_html = add_content_type_charset(ContentType.TEXT_HTML, "UTF-8")
         ctype_xml = add_content_type_charset(ContentType.APP_XML, "UTF-8")
         proc_url = process.href(request)
         if ctype in ContentType.ANY_XML or str(schema).upper() == ProcessSchema.WPS:
             offering = process.offering(ProcessSchema.WPS, request=request)
             headers = [
-                ("Link", f"<{proc_url}?f=json>; rel=\"alternate\"; type={ctype_json}"),
-                ("Link", f"<{proc_url}?f=html>; rel=\"alternate\"; type={ctype_html}"),
+                ("Link", make_link_header(f"{proc_url}?f=json", rel="alternate", type=ctype_json)),
+                ("Link", make_link_header(f"{proc_url}?f=html", rel="alternate", type=ctype_html)),
+                ("Link", make_link_header(f"{proc_url}?f=yaml", rel="alternate", type=ctype_yaml)),
                 ("Content-Type", ctype_xml),
             ]
             return Response(offering, headerlist=headers)
-        else:
+        elif ctype == ContentType.APP_YAML:
             offering = process.offering(schema)
-            fmt_alt, ctype_alt = ("html", ctype_html) if ctype == ContentType.APP_JSON else ("json", ctype_json)
+            content = OutputFormat.convert(offering, OutputFormat.YAML)
+            headers = [
+                ("Link", make_link_header(f"{proc_url}?f=json", rel="alternate", type=ctype_json)),
+                ("Link", make_link_header(f"{proc_url}?f=html", rel="alternate", type=ctype_html)),
+                ("Link", make_link_header(f"{proc_url}?f=xml", rel="alternate", type=ctype_xml)),
+                ("Link", make_link_header(sd.OGC_API_PROC_PROFILE_PROC_DESC_URL, rel="profile")),
+                ("Content-Profile", sd.OGC_API_PROC_PROFILE_PROC_DESC_URL),
+                ("Content-Type", ctype_yaml),
+            ]
+            return HTTPOk(headers=headers, content_type=ctype, charset="utf-8", body=content)
+        elif ctype == ContentType.APP_JSON:
+            offering = process.offering(schema)
+            request.response.content_type = ctype_json
             request.response.headers.extend([
-                ("Link", f"<{proc_url}?f=xml>; rel=\"alternate\"; type={ctype_xml}"),
-                ("Link", f"<{proc_url}?f={fmt_alt}>; rel=\"alternate\"; type={ctype_alt}")
+                ("Link", make_link_header(f"{proc_url}?f=xml", rel="alternate", type=ctype_xml)),
+                ("Link", make_link_header(f"{proc_url}?f=yaml", rel="alternate", type=ctype_yaml)),
+                ("Link", make_link_header(f"{proc_url}?f=html", rel="alternate", type=ctype_html)),
+                ("Link", make_link_header(sd.OGC_API_PROC_PROFILE_PROC_DESC_URL, rel="profile")),
+                ("Content-Profile", sd.OGC_API_PROC_PROFILE_PROC_DESC_URL),
+            ])
+            return Box(offering)
+        else:  # HTML
+            offering = process.offering(schema)
+            request.response.headers.extend([
+                ("Link", make_link_header(f"{proc_url}?f=json", rel="alternate", type=ctype_json)),
+                ("Link", make_link_header(f"{proc_url}?f=yaml", rel="alternate", type=ctype_yaml)),
+                ("Link", make_link_header(f"{proc_url}?f=xml", rel="alternate", type=ctype_xml)),
             ])
             return Box(offering)
     # FIXME: handle colander invalid directly in tween (https://github.com/crim-ca/weaver/issues/112)
@@ -303,7 +332,14 @@ def get_local_process(request):
 @sd.process_package_service.get(
     tags=[sd.TAG_PROCESSES, sd.TAG_DESCRIBEPROCESS],
     schema=sd.ProcessPackageEndpoint(),
-    accept=ContentType.APP_JSON,
+    accept=[
+        ContentType.APP_JSON,
+        ContentType.APP_YAML,
+        ContentType.APP_CWL,
+        ContentType.APP_CWL_JSON,
+        ContentType.APP_CWL_YAML,
+        ContentType.APP_CWL_X,
+    ],
     renderer=OutputFormat.JSON,
     response_schemas=sd.get_process_package_responses,
 )
@@ -314,7 +350,22 @@ def get_local_process_package(request):
     Get a registered local process package definition.
     """
     process = get_process(request=request)
-    return HTTPOk(json=process.package or {})
+    content_type = get_header("Accept", request.headers, default=ContentType.APP_CWL_JSON)
+    # ignore default browser request injecting HTML
+    # ignore 'weaver.wps_restapi_html_override_user_agent' as well since HTML cannot apply here
+    if all(ctype in content_type for ctype in [ContentType.TEXT_HTML, ContentType.ANY]):
+        content_type = ContentType.APP_CWL_JSON
+    headers = {
+        "Link": make_link_header(sd.CWL_SCHEMA_URL, rel="profile", type=ContentType.APP_YAML),
+        "Content-Schema": sd.CWL_SCHEMA_URL,
+        "Content-Profile": sd.CWL_SCHEMA_URL,
+    }
+    yml_fmt = [ContentType.APP_YAML, ContentType.APP_CWL_YAML]
+    cwl_fmt = OutputFormat.YAML if any(ctype in content_type for ctype in yml_fmt) else OutputFormat.JSON
+    package = OutputFormat.convert(process.package, cwl_fmt)
+    content = {"json": package} if cwl_fmt == OutputFormat.JSON else {"body": package}
+    content = content if package else {}
+    return HTTPOk(headers=headers, content_type=content_type, charset="utf-8", **content)
 
 
 @sd.process_payload_service.get(